Add path_for helper
path_for is like url_for with only_path always set to true. This makes
it easy to generate proper paths in a child controller
(EmbeddedQuizQuestionsController) even though the code that generates
them is in the parent controller (QuizQuestionsController) by not
specifying the controller (and thus current controller being used to
generate the route).